Atualizações de compatibilidade e loading em envio de e-mail com relatório de visitas do dia.

+++ PropagaMed/PropagaMed/View/Home.xaml.cs
@@ -2,6 +2,7 @@
 using PropagaMed.View;
 using System;
 using System.Collections.Generic;
+using System.ComponentModel;
 using System.IO;
 using System.Linq;
 using System.Net;
@@ -20,13 +21,9 @@ namespace PropagaMed
         public Home(bool otherView = false)
         {
             InitializeComponent();
-
-            // Binding para data mínima ser a atual
-            DateTime dateDeHoje = DateTime.Now.Date;
-            dataVisita.BindingContext = dateDeHoje;
-            exportToCSV.Text = $"Exportar Visitas do Dia {DateTime.Now:dd/MM/yyyy}";
-
             AlimentaMedicosEVisitas();
+
+            exportToCSV.Text = $"Exportar Visitas do Dia {DateTime.Now:dd/MM/yyyy}";
 
             this.CurrentPage = otherView ? verMedicos : this.CurrentPage;
         }
@@ -64,7 +61,6 @@ namespace PropagaMed
             if (!String.IsNullOrEmpty(MedicoSelecionado.Nome))
             {
                 await App.Database.SaveItemAsync(VisitaASalvar);
-                await DisplayAlert("Informação", "Visita para médico(a) " + MedicoSelecionado.Nome + " às " + horaVisita.Time.ToString(@"hh\:mm") + " em " + DataSelecionada.ToString("dd/MM/yyyy") + " cadastrada com sucesso", "Ok");
                 AlimentaMedicosEVisitas();
                 this.CurrentPage = verVisitas;
             }
@@ -114,7 +110,6 @@ namespace PropagaMed
             if (!String.IsNullOrEmpty(MedicoASalvar.Nome))
             {
                 await App.Database.SaveItemAsync(MedicoASalvar);
-                await DisplayAlert("Informação", "Médico(a) " + MedicoASalvar.Nome + " cadastrado(a) com sucesso", "Ok");
                 AlimentaMedicosEVisitas();
                 this.CurrentPage = novaVisita;
             }
@@ -191,6 +186,9 @@ namespace PropagaMed
 
         async void ExportToCSV(object sender, EventArgs e)
         {
+            actInd.IsRunning = true;
+            await Task.Delay(100);
+
             var title = $"PropagaMed - Relatório de Controle de Visitas - {DateTime.Now:dd/MM/yyyy}";
             string to = "luidi.lima@poli.ufrj.br";
             string from = "luidi.lima@poli.ufrj.br";
@@ -198,42 +196,51 @@ namespace PropagaMed
             List<Visita> visitas = App.Database.GetItemsVisitaDiaAsync().Result.OrderBy(v => v.HoraVisita).ToList();
             List<Medico> medicos = new List<Medico>();
 
-            Parallel.ForEach(visitas, visita =>
+            if (visitas.Any())
             {
-                medicos.Add(App.Database.GetItemAsync(visita.IdMedicoVisita).Result);
-            });
+                Parallel.ForEach(visitas, visita =>
+                {
+                    medicos.Add(App.Database.GetItemAsync(visita.IdMedicoVisita).Result);
+                });
 
-            if (!File.Exists(personalFolderFile))
+                if (!File.Exists(personalFolderFile))
+                {
+                    var content = new List<string>() { "Controle de Visitas", "Nº;CRM;Nome;Data;Hora;Especialidade;Observação;" };
+                    var toAdd = visitas.Select(v => string.Join(";", visitas.IndexOf(v) + 1, medicos.Where(m => m.Id.Equals(v.IdMedicoVisita)).First()?.CRM, v.NomeMedicoVisita, v.DiaVisita.ToString("dd/MM/yyyy"), v.HoraVisita.ToString(@"hh\:mm"), medicos.Where(m => m.Id.Equals(v.IdMedicoVisita)).First()?.Especialidade, v.Observacao)).ToList();
+
+                    content.AddRange(toAdd);
+
+                    File.WriteAllLines(personalFolderFile, content.ToArray(), System.Text.Encoding.UTF8);
+                }
+
+                SmtpClient smtp = new SmtpClient("smtp.gmail.com", 587);
+                smtp.Credentials = new NetworkCredential("luidi.lima@poli.ufrj.br", string.Empty/*SENHA AQUI*/);
+                smtp.DeliveryMethod = SmtpDeliveryMethod.Network;
+                smtp.EnableSsl = true;
+
+                MailMessage mailSend = new MailMessage();
+
+                mailSend.From = new MailAddress(from.Replace(';', ','));
+                mailSend.To.Add(to.Replace(';', ','));
+                mailSend.Subject = title;
+                mailSend.SubjectEncoding = System.Text.Encoding.UTF8;
+                mailSend.Body = "Anexo.";
+                mailSend.BodyEncoding = System.Text.Encoding.UTF8;
+                mailSend.IsBodyHtml = true;
+                mailSend.Attachments.Add(new Attachment(personalFolderFile, MediaTypeNames.Application.Octet));
+                smtp.Send(mailSend);
+
+                if (File.Exists(personalFolderFile))
+                    File.Delete(personalFolderFile);
+
+                await DisplayAlert("Informação", $"Arquivo enviado por e-mail", "Ok");
+            }
+            else
             {
-                var content = new List<string>() { "Controle de Visitas", "Nº;CRM;Nome;Data;Hora;Especialidade;Observação;" };
-                var toAdd = visitas.Select(v => string.Join(";", visitas.IndexOf(v)+1, medicos.Where(m => m.Id.Equals(v.IdMedicoVisita)).First()?.CRM, v.NomeMedicoVisita, v.DiaVisita.ToString("dd/MM/yyyy"), v.HoraVisita.ToString(@"hh\:mm"), medicos.Where(m => m.Id.Equals(v.IdMedicoVisita)).First()?.Especialidade, v.Observacao)).ToList();
-
-                content.AddRange(toAdd);
-
-                File.WriteAllLines(personalFolderFile, content.ToArray(), System.Text.Encoding.UTF8);
+                await DisplayAlert("Informação", $"Não há visitas para {DateTime.Now:dd/MM/yyyy}", "Ok");
             }
 
-            SmtpClient smtp = new SmtpClient("smtp.gmail.com", 587);
-            smtp.Credentials = new NetworkCredential("luidi.lima@poli.ufrj.br", string.Empty/*SENHA AQUI*/);
-            smtp.DeliveryMethod = SmtpDeliveryMethod.Network;
-            smtp.EnableSsl = true;
-
-            MailMessage mailSend = new MailMessage();
-
-            mailSend.From = new MailAddress(from.Replace(';', ','));
-            mailSend.To.Add(to.Replace(';', ','));
-            mailSend.Subject = title;
-            mailSend.SubjectEncoding = System.Text.Encoding.UTF8;
-            mailSend.Body = "Anexo.";
-            mailSend.BodyEncoding = System.Text.Encoding.UTF8;
-            mailSend.IsBodyHtml = true;
-            mailSend.Attachments.Add(new Attachment(personalFolderFile, MediaTypeNames.Application.Octet));
-            smtp.Send(mailSend);
-
-            if (File.Exists(personalFolderFile))
-                File.Delete(personalFolderFile);
-
-            await DisplayAlert("Informação", $"Arquivo enviado por e-mail", "Ok");
+            actInd.IsRunning = false;
         }
     }
 }
